feat: Seed data script for DB data

+++ 06-06-2025/QuizupAPI/Models/DTOs/StudentSummaryDTO.cs
@@ -1,4 +1,5 @@
 using System.Text.Json;
+using System.ComponentModel.DataAnnotations;
 
 namespace QuizupAPI.Models.DTOs.Student
 {
@@ -8,16 +9,16 @@ namespace QuizupAPI.Models.DTOs.Student
         public string StudentName { get; set; } = string.Empty;
         public string StudentEmail { get; set; } = string.Empty;
         public string StudentClass { get; set; } = string.Empty;
-        public int TotalQuizzesAvailable { get; set; }
-        public int TotalQuizzesStarted { get; set; }
-        public int TotalQuizzesCompleted { get; set; }
-        public int TotalQuizzesInProgress { get; set; }
-        public int TotalQuizzesSaved { get; set; }
+        public long TotalQuizzesAvailable { get; set; }
+        public long TotalQuizzesStarted { get; set; }
+        public long TotalQuizzesCompleted { get; set; }
+        public long TotalQuizzesInProgress { get; set; }
+        public long TotalQuizzesSaved { get; set; }
         public decimal AverageScore { get; set; }
         public int HighestScore { get; set; }
         public int LowestScore { get; set; }
-        public int TotalQuestionsAttempted { get; set; }
-        public int TotalCorrectAnswers { get; set; }
+        public long TotalQuestionsAttempted { get; set; }
+        public long TotalCorrectAnswers { get; set; }
         public decimal AccuracyPercentage { get; set; }
         public int TotalTimeSpentMinutes { get; set; }
         public JsonDocument QuizzesByStatus { get; set; } = JsonDocument.Parse("{}");
