+++ 06-06-2025/QuizupAPI/Models/DTOs/TeacherSummaryDTO.cs
@@ -1,4 +1,5 @@
 using System.Text.Json;
+using System.ComponentModel.DataAnnotations;
 
 namespace QuizupAPI.Models.DTOs.Teacher
 {
@@ -8,18 +9,18 @@ namespace QuizupAPI.Models.DTOs.Teacher
         public string TeacherName { get; set; } = string.Empty;
         public string TeacherEmail { get; set; } = string.Empty;
         public string TeacherSubject { get; set; } = string.Empty;
-        public int TotalQuizzesCreated { get; set; }
-        public int TotalActiveQuizzes { get; set; }
-        public int TotalInactiveQuizzes { get; set; }
-        public int TotalQuestionsCreated { get; set; }
-        public int TotalStudentSubmissions { get; set; }
-        public int TotalStudentsParticipated { get; set; }
+        public long TotalQuizzesCreated { get; set; }
+        public long TotalActiveQuizzes { get; set; }
+        public long TotalInactiveQuizzes { get; set; }
+        public long TotalQuestionsCreated { get; set; }
+        public long TotalStudentSubmissions { get; set; }
+        public long TotalStudentsParticipated { get; set; }
         public decimal AverageCompletionRate { get; set; }
         public decimal AverageStudentScore { get; set; }
         public int HighestQuizScore { get; set; }
         public int LowestQuizScore { get; set; }
-        public int TotalQuestionsAnswered { get; set; }
-        public int TotalCorrectAnswers { get; set; }
+        public long TotalQuestionsAnswered { get; set; }
+        public long TotalCorrectAnswers { get; set; }
         public decimal OverallAccuracyPercentage { get; set; }
         public JsonDocument QuizzesByStatus { get; set; } = JsonDocument.Parse("{}");
         public JsonDocument StudentPerformanceSummary { get; set; } = JsonDocument.Parse("[]");
